+++ C#/Utils/MathUtils.cs
@@ -5,11 +5,11 @@ namespace EulerProblem.Utils
 {
     public static class MathUtils
     {
-        public static int GCD(int number1, int number2)
+        public static BigInteger GCD(BigInteger number1, BigInteger number2)
         {
             while (number1 % number2 != 0)
             {
-                int temp = number2;
+                BigInteger temp = number2;
                 number2 = number1 % number2;
                 number1 = temp;
             }
